Major code refactoring

+++ classifier/classes/modules/images/images_cnn/ModelImagesCNN.py
@@ -7,4 +7,4 @@ class ModelImagesCNN(Model):
 
     def __init__(self, network_params: dict):
         super().__init__(device=network_params["device"])
-        self._network = ImagesCNN(network_params).float().to(self._device)
+        self._network = ImagesCNN(network_params).to(self._device)
