+++ classifier/classes/modules/images/images_cnn_rnn/ModelImagesCNNRNN.py
@@ -6,4 +6,4 @@ class ModelImagesCNNRNN(Model):
 
     def __init__(self, network_params: dict):
         super().__init__(device=network_params["device"])
-        self._network = ImagesCNNRNN(network_params).float().to(self._device)
+        self._network = ImagesCNNRNN(network_params).to(self._device)
