+++ classifier/classes/modules/images/pre_trained_cnn/ModelPreTrainedCNN.py
@@ -6,4 +6,4 @@ class ModelPreTrainedCNN(Model):
 
     def __init__(self, network_params: dict):
         super().__init__(device=network_params["device"])
-        self._network = PreTrainedCNN(network_params).float().to(self._device)
+        self._network = PreTrainedCNN(network_params).to(self._device)
