+++ classifier/classes/modules/sequences/sequences_cnn/ModelSequencesCNN.py
@@ -7,4 +7,4 @@ class ModelSequencesCNN(Model):
 
     def __init__(self, network_params: dict):
         super().__init__(device=network_params["device"])
-        self._network = SequencesCNN(network_params).float().to(self._device)
+        self._network = SequencesCNN(network_params).to(self._device)
